+++ R/sfReapportion.R
@@ -15,12 +15,12 @@
 #' @param weight_matrix \strong{(not yet supported)} a `SpatialPointsDataFrame` indicating where are the observations (inhabitants, voters, etc.) (optional).
 #' @param weight_matrix_var \strong{(not yet supported)} the name of the variable in \code{weight_matrix} containing the weights.
 #' @export
-#' @import sp sf purrr
+#' @import sp sf
+#' @importFrom dplyr group_by
 #' @importFrom dplyr left_join
-#' @importFrom dplyr mutate_if
-#' @importFrom dplyr pull
-#' @importFrom plyr ddply
-#' @importFrom plyr numcolwise
+#' @importFrom dplyr mutate_at
+#' @importFrom dplyr select
+#' @importFrom dplyr summarise_at
 #' @importFrom purrr map_int
 #'
 sfReapportion <- function(old_geom, new_geom, data, old_ID, new_ID, data_ID, variables = names(data)[-which(names(data) %in% data_ID)], mode = "count", weights = NULL, weight_matrix = NULL, weight_matrix_var = NULL) {
@@ -34,7 +34,19 @@ sfReapportion <- function(old_geom, new_geom, data, old_ID, new_ID, data_ID, var
 
   if (!(old_ID %in% names(old_geom@data))) stop(paste(old_ID, "is not a variable from", deparse(substitute(old_geom)),"!", sep=" "))
   if (!(new_ID %in% names(new_geom@data))) stop(paste(new_ID, "is not a variable from", deparse(substitute(new_geom)),"!", sep=" "))
-  if (sum(!(variables %in% names(data))) > 0) stop(paste(variables[!(variables %in% names(data))], "is not a variable from", deparse(substitute(data)),"!",sep=" "))
+
+  # check variables
+  if (sum(!(variables %in% names(data))) > 0)
+    stop(paste(variables[!(variables %in% names(data))], "is not a variable from", deparse(substitute(data)),"!",sep=" "))
+  ###
+  ### exclude non-numeric variables from reapportionment
+  ###
+  excl <- base::setdiff(variables, names(Filter(is.numeric, data[, variables])))
+  if (length(excl) > 0) {
+    variables <- variables[ !variables %in% excl ]
+    warning("non-numeric variable(s) ignored: ", paste(excl, collapse = ", "))
+  }
+
   if (mode %in% "proportion" & is.null(weights)) stop("When mode = 'proportion', you must provide weights.")
   if (mode %in% "proportion")
     if (!(weights %in% names(data)))
@@ -109,7 +121,7 @@ sfReapportion <- function(old_geom, new_geom, data, old_ID, new_ID, data_ID, var
                          sf::st_drop_geometry(int)[, new_ID ])
 
   ###
-  ### crucial difference with `rgeos` -- read names from row names, not from `names`
+  ### crucial difference with {rgeos} -- get IDs from `rownames` instead of `names`
   ###
   # intdf <- data.frame(intname = names(int)) # make a data frame for the intersected SpatialPolygon, using names from the output list from int
   intdf <- data.frame(intname = rownames(int)) # make a data frame for the intersected SpatialPolygon, using names from the output list from int
@@ -141,17 +153,27 @@ sfReapportion <- function(old_geom, new_geom, data, old_ID, new_ID, data_ID, var
     ### switch to `st_area` (again)
     ###
     # data$departarea <- rgeos::gArea(old_geom, byid = TRUE)[match(data$old_ID, old_geom@data[, old_ID])]
-    data$departarea <- sf::st_area(old_geom)[match(data$old_ID, dplyr::pull(sf::st_drop_geometry(old_geom[, old_ID])))]
+    data$departarea <- sf::st_area(old_geom)[ match(data$old_ID, old_geom[[ old_ID]]) ]
   }
 
+  # note: `left_join` is much faster than `base::merge`
   intdf2 <- dplyr::left_join(intdf, data, by = c("old_ID" = "old_ID")) # join together the two dataframes by the administrative ID
   if (mode %in% "count") {
-    intdf2[,paste(variables,"inpoly",sep="")] <- plyr::numcolwise(function(x) {x * (intdf2$polyarea / intdf2$departarea)})(as.data.frame(intdf2[,variables]))
-    intpop <- plyr::ddply(intdf2, "new_ID", function(x) {plyr::numcolwise(sum, na.rm = TRUE)(as.data.frame(x[,paste(variables,"inpoly",sep="")]))}) # sum population lying within each polygon
-    names(intpop)[-1] <- variables
+    ###
+    ### note: {dplyr} code roughly 1.5x faster than older {plyr} code
+    ###
+    intpop <- dplyr::mutate_at(intdf2, variables, ~ .x * (intdf2$polyarea / intdf2$departarea))
+    # remove `units` type (drastically speeds up the `sum` operation below)
+    intpop <- dplyr::mutate_at(intpop, variables, as.double)
+    # subset to target variables
+    intpop <- dplyr::select(intpop, new_ID, variables)
+    # aggregate by sum (as Joël wrote, other functions might be suitable)
+    intpop <- dplyr::group_by(intpop, new_ID)
+    intpop <- dplyr::summarise_at(intpop, variables, sum, na.rm = TRUE)
   } else {
     stop("use of weight matrix not yet supported")
-    # intdf2[,paste(variables,"inpoly",sep="")] <- plyr::numcolwise(function(x) {x * (intdf2$polyarea / intdf2$departarea)})(as.data.frame(intdf2[,variables]) * intdf2[,weights])
+    # intdf2[,paste(variables,"inpoly",sep="")] <- dplyr::mutate_all(as.data.frame(intdf2[,variables]),
+    #                                                                ~ .x * (intdf2$polyarea / intdf2$departarea))
     # intdf2$weights <- intdf2[,weights] * (intdf2$polyarea / intdf2$departarea)
     # intpop <- plyr::ddply(intdf2, "new_ID", function(x) {plyr::numcolwise(sum, na.rm = TRUE)(as.data.frame(x[,c(paste(variables,"inpoly",sep=""), "weights")]))}) # sum population lying within each polygon
     # intpop[,paste0(variables, "inpoly")] <- intpop[,paste0(variables, "inpoly")] / intpop$weights
@@ -159,14 +181,13 @@ sfReapportion <- function(old_geom, new_geom, data, old_ID, new_ID, data_ID, var
     # names(intpop)[length(names(intpop))] <- weights
   }
 
-
-
   names(intpop)[1] <- new_ID
 
   ###
-  ### remove `units` type
+  ### return strict data.frame, for compatibility with output of {spReapportion}
   ###
-  intpop <- dplyr::mutate_if(intpop, is.numeric, as.double)
+  return(as.data.frame(intpop)) # done!
 
-  return(intpop) # done!
 }
+
+# kthxbye