+++ NEWS.md
@@ -1,3 +1,9 @@
+# version 0.1.1
+
+- additional check on `variables` (numeric only)
+- allowing `sf` objects as inputs
+- losing dependency to `plyr` (faster code, too)
+
 # version 0.1.0
 
 - `sfReapportion` seems to work without calling `rgeos` or `maptools`